fixed some mistakes I did during rebasing
