+++ frontend/script.js
@@ -1,11 +1,26 @@
 const express = require("express");
 const path = require("path");
 const app = express();
+const fetch = require("node-fetch");
+
+const coreAPI = "http://core:8080/data"; // this is the exposed API from core container
+
+let dataSet;
+
+async function fetchData() {
+  const response = await fetch(coreAPI);
+  await response.json().then((data) => (dataSet = data));
+  return response;
+}
 
 app.use(express.static(__dirname + "/public"));
 
 app.get("/", function (req, res) {
   res.sendFile(path.join(__dirname + "/index.html"));
+});
+
+app.get("/core", (req, res) => {
+  fetchData().then(() => res.send(dataSet));
 });
 
 app.listen(8080);
